processor/requirements.txt
Pillow>=10.0,<12
PyMuPDF>=1.24,<2
reportlab>=4.0,<5
